+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Suppress Spotify API HTTP errors from flooding console output
 * Change default volume from 80% to 100%
 * Require Last.fm to be configured and working to start
+* Fix volume not persisting when track changes
 
 ### 1.0.10: 2026-01-07